refactor(seshat): split plugin into modules + fix ButtonSize casing, add sidebar "+"
Bug fix (caught at runtime): icon-button `button-size` deserializes as
`Small`/`Medium`/`Large` (PascalCase), not lowercase — the sidebar's `"small"`
(and the editor's `"medium"`) blanked the render. Corrected both.

Structure: lib.rs was a single ~620-line file; split into focused modules:
  - state.rs   — Connection/Form/State, engine helpers, persistence, Request
  - events.rs  — apply_event + connection actions + query-result folding
  - ui/mod.rs  — connections manager, SQL editor, new-connection dialog
  - ui/sidebar.rs — the sidebar view
  - lib.rs     — WIT guest impls + shared helpers + export! only

Also adds a "+" New-connection button to the sidebar header. The sidebar runs
as its own wasm instance and the host only re-renders the sidebar after a
sidebar event, so it carries its own copy of the modal.

The render-parse test now also renders the sidebar and the editor view and
parses both as host UiNode — exactly the regression class (bad enum casing)
that slipped through before. clippy (host + wasm) + fmt clean; 197 tests pass.
